flask_app/requirements.txt
# Zależności Python dla aplikacji Flask
flask
htmx
pyyaml
